+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.mocked_api import MockedAPI
-from wbapi_async.client.session import BaseSession
+from wbapi.client.session import BaseSession
 
 
 @pytest.fixture